+++ lib/src/widget/control_button_bar.dart
@@ -1,6 +1,7 @@
+import 'package:filter_list/filter_list.dart';
 import 'package:filter_list/src/state/filter_state.dart';
 import 'package:filter_list/src/state/provider.dart';
-import 'package:filter_list/src/theme/theme.dart';
+
 import 'package:filter_list/src/widget/control_button.dart';
 import 'package:flutter/material.dart';
 
@@ -12,12 +13,16 @@ class ControlButtonBar<T> extends StatelessWidget {
     this.resetButtonText,
     this.applyButtonText,
     this.onApplyButtonClick,
+    required this.controlButtons,
   }) : super(key: key);
   final bool enableOnlySingleSelection;
   final String? allButtonText;
   final String? resetButtonText;
   final String? applyButtonText;
   final VoidCallback? onApplyButtonClick;
+
+  /// {@macro control_buttons}
+  final List<ContolButtonType> controlButtons;
 
   @override
   Widget build(BuildContext context) {
@@ -40,7 +45,8 @@ class ControlButtonBar<T> extends StatelessWidget {
                   mainAxisSize: MainAxisSize.min,
                   children: <Widget>[
                     /* All Button */
-                    if (!enableOnlySingleSelection) ...[
+                    if (!enableOnlySingleSelection &&
+                        controlButtons.contains(ContolButtonType.All)) ...[
                       ContorlButton(
                         choiceChipLabel: '$allButtonText',
                         onPressed: () {
@@ -55,18 +61,19 @@ class ControlButtonBar<T> extends StatelessWidget {
 
                       /* Reset Button */
                     ],
-                    ContorlButton(
-                      choiceChipLabel: '$resetButtonText',
-                      onPressed: () {
-                        final state = StateProvider.of<FilterState<T>>(
-                          context,
-                          rebuildOnChange: true,
-                        );
-                        state.selctedItems = [];
-                      },
-                    ),
-                    SizedBox(width: theme.buttonSpacing),
-
+                    if (controlButtons.contains(ContolButtonType.Reset)) ...[
+                      ContorlButton(
+                        choiceChipLabel: '$resetButtonText',
+                        onPressed: () {
+                          final state = StateProvider.of<FilterState<T>>(
+                            context,
+                            rebuildOnChange: true,
+                          );
+                          state.selctedItems = [];
+                        },
+                      ),
+                      SizedBox(width: theme.buttonSpacing),
+                    ],
                     /* Apply Button */
                     ContorlButton(
                       choiceChipLabel: '$applyButtonText',