+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 - Add `HeaderTheme` for Header widget theme
 - Add `ControlButtonBarTheme` for control button bar theme
 - Add `ControlButtonTheme` for control button theme
+- Add `controlButtons` option to display/hide control buttons (All, Reset)
 
 ## [1.0.0] 21 May 2021